fix: DropRateMaster

+++ Assets/PrimeMillionaire/Scripts/Common/Data/DataStore/DropRateMaster.cs
@@ -6,14 +6,14 @@ namespace PrimeMillionaire.Common.Data.DataStore
     [MemoryTable(nameof(DropRateMaster)), MessagePackObject(true)]
     public sealed class DropRateMaster
     {
-        public DropRateMaster(int turn, float rate)
+        public DropRateMaster(int turn, int rate)
         {
             Turn = turn;
             Rate = rate;
         }
 
         [PrimaryKey] public int Turn { get; }
-        public float Rate { get; }
+        public int Rate { get; }
 
         public DropRateVO ToVO() => new(Turn, Rate);
     }
